zad_rest/requirements.txt
anyio==3.6.2
click==8.1.3
Flask==2.2.3
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.6
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
Werkzeug==2.2.3
zipp==3.15.0
requests~=2.28.2
python-dotenv~=1.0.0